refactor: move regulatory compliance constants to a dedicated file for better organization
