+++ plugins/wazuh-analysis/public/groups/security-operations/applications.ts
@@ -5,14 +5,14 @@ import {
   AppUpdater,
 } from '../../../../../src/core/public';
 import {
-  REGULATORY_COMPLIANCE_ID,
-  REGULATORY_COMPLIANCE_TITLE,
-} from './constants';
-import {
   INCIDENT_RESPONSE_ID,
   INCIDENT_RESPONSE_TITLE,
 } from './apps/incident-response/constants';
 import { IT_HYGIENE_ID, IT_HYGIENE_TITLE } from './apps/it-hygiene/constants';
+import {
+  REGULATORY_COMPLIANCE_ID,
+  REGULATORY_COMPLIANCE_TITLE,
+} from './apps/regulatory-compliance/constants';
 
 export function getSecurityOperationsApps(updater$?: Subject<AppUpdater>) {
   return [
