+++ plugins/wazuh-analysis/public/groups/security-operations/apps/regulatory-compliance/regulatory-compliance-app.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { AppMountParameters } from 'opensearch-dashboards/public';
-import { REGULATORY_COMPLIANCE_TITLE } from '../../constants';
+import { REGULATORY_COMPLIANCE_TITLE } from './constants';
 
 interface RegulatoryComplianceAppProps {
   params: AppMountParameters;
